requirements.txt
streamlit
ultralytics
torch
numpy
pillow
opencv-python